backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-openai
langchain-community
pgvector
psycopg2-binary
pdf2image
openai
pydantic-settings
tavily-python
python-multipart
